.build_and_deploy/pyproject.toml
[tool.poetry]
package-mode = false
license ="Apache-2.0"
description = "My own personal website. Available at https://carrotmanmatt.com"
authors = ["Matt Norton <matt@carrotmanmatt.com>"]
repository = "https://github.com/CarrotManMatt/website.git"
keywords = ["website", "html", "HTML5up", "carrot"]
classifiers = [
    "Private :: Do Not Upload",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
    "Topic :: Internet",
]

[tool.poetry.dependencies]
python = "^3.12"
django = "~5.0"
classproperties = {git = "https://github.com/hottwaj/classproperties.git"}
setuptools = "^69.5"
gitpython = "^3.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7"
mypy = "~1.10"
django-stubs = {extras = ["compatible-mypy"], version = "~5.0"}
ruff = "^0.5"
djlint = "^1.34"


[tool.mypy]
plugins = ["mypy_django_plugin.main"]
exclude = ["venv/"]
extra_checks = true
warn_unused_configs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_subclassing_any = true
disallow_untyped_decorators = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
warn_unreachable = true


[tool.ruff]
line-length = 95
target-version ="py312"
extend-exclude = [
    "venv/",
    ".vscode",
    ".idea",
    ".mypy_cache",
    ".ruff_cache",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "C",
    "I",
    "N",
    "D",
    "UP",
    "YTT",
    "ANN",
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "DTZ",
    "T",
    "DJ",
    "EM",
    "G",
    "INP",
    "PIE",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "ARG",
    "PTH",
    "TD",
    "FIX",
    "ERA",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "PERF",
    "RUF",
]
fixable = ["I001", "TCH", "F401", "COM812"]
ignore = [
    "N818",
    "N806",
    "D203",
    "D212",
    "ANN101",
    "ANN102",
    "Q003",
    "TD002",
    "TD003",
    "S311",
    "UP040",  # NOTE: Mypy does not currently support PEP 695 type aliases, so they should not be used
    "PT009",
    "PT027",
]
task-tags = [
    "TODO",
    "FIXME",
    "HACK",
    "BUG",
    "NOBUG",
    "REQ",
    "IDEA",
    "NOTE",
    "ISSUE",
    "DONE",
]

[tool.ruff.lint.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true
max-doc-length = 95

[tool.ruff.lint.pylint]
allow-magic-value-types = ["str", "bytes", "int"]
max-args = 7
max-returns = 10
max-branches = 19

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.djlint]
profile = "django"
use_gitignore = true
include = "H017,H035"
blank_line_after_tag="load,extends,include"
